Refactorizar para condicionar si el usuario existe en las peticiones

+++ 07_Backend/06_queries-knex/models/UsersModel.js
@@ -11,19 +11,33 @@ const getAllUsers = () => {
   return db
     .select('*')
     .from('users')
-    .where({ is_active: true })
-    .then(result => result);
+    .where({ is_active: true });
 };
 
-const getUserById = id => {
-  return db
+const getUserById = async id => {
+  const user = await db
     .select('*')
     .from('users')
     .where({ user_id: id, is_active: true })
     .then(result => result[0]);
+
+  if (!user) {
+    throw new Error('El usuario que quieres traer no existe');
+  }
+  return user;
 };
 
-const updateUser = (id, body) => {
+const updateUser = async (id, body) => {
+  const user = await db
+    .select('*')
+    .from('users')
+    .where({ user_id: id, is_active: true })
+    .then(result => result[0]);
+
+  if (!user) {
+    throw new Error('El usuario que quieres actualizar no existe');
+  }
+
   return db('users')
     .update(body)
     .where({ user_id: id })
@@ -31,7 +45,17 @@ const updateUser = (id, body) => {
     .then(result => result[0]);
 };
 
-const deleteUser = id => {
+const deleteUser = async id => {
+  const user = await db
+    .select('*')
+    .from('users')
+    .where({ user_id: id, is_active: true })
+    .then(result => result[0]);
+
+  if (!user) {
+    throw new Error('El usuario que quieres eliminar no existe');
+  }
+
   return db('users')
     .update({ is_active: false })
     .where({ user_id: id })
@@ -39,7 +63,18 @@ const deleteUser = id => {
     .then(result => result[0]);
 };
 
-const destroyUser = id => {
+const destroyUser = async id => {
+  const user = await db
+    .select('*')
+    .from('users')
+    .where({ user_id: id, is_active: false })
+    .then(result => result[0]);
+
+  if (!user) {
+    throw new Error(
+      'El usuario que quieres destruir aún se encuentra activo o no existe'
+    );
+  }
   return db('users')
     .where({ user_id: id, is_active: false })
     .delete();
